+++ src/packages/sidenavbaritem/index.ts
@@ -1,4 +1,0 @@
-import { SideNavBarItem } from './sidenavbaritem'
-
-export type { SideNavBarItemProps } from './sidenavbaritem'
-export default SideNavBarItem